+++ node/resolvers/index.ts
@@ -181,10 +181,6 @@ export const resolvers = {
               organizationResponse?.data?.getOrganizationById?.collections
                 ?.length
             ) {
-              console.log(
-                'Collection =>',
-                organizationResponse.data.getOrganizationById.collections
-              )
               const collections =
                 organizationResponse.data.getOrganizationById.collections.map(
                   (collection: any) => `productClusterIds=${collection.id}`