+++ node/roleNames.ts
@@ -4,9 +4,9 @@ const roleNames: any = {
     'sales-admin': 'Sales Admin',
     'sales-manager': 'Sales Manager',
     'sales-representative': 'Sales Representative',
-    'customer-admin': 'Admin',
-    'customer-approver': 'Approver',
-    'customer-buyer': 'Buyer',
+    'customer-admin': 'Organization Admin',
+    'customer-approver': 'Organization Approver',
+    'customer-buyer': 'Organization Buyer',
   },
 }